+++ internal/pkg/chainsync/internal/syncer/syncer_test.go
@@ -381,7 +381,7 @@ func TestBlocksNotATipSetRejected(t *testing.T) {
 	b2 := builder.AppendBlockOnBlocks(b1)
 
 	badKey := block.NewTipSetKey(b1.Cid(), b2.Cid())
-	err := syncer.HandleNewTipSet(ctx, block.NewChainInfo(peer.ID(""), "", badKey, uint64(b1.Height)), false)
+	err := syncer.HandleNewTipSet(ctx, block.NewChainInfo(peer.ID(""), "", badKey, b1.Height), false)
 	assert.Error(t, err)
 
 	_, err = store.GetTipSet(badKey)
@@ -421,14 +421,14 @@ func newPoisonValidator(t *testing.T, headerFailure, fullFailure uint64) *poison
 
 func (pv *poisonValidator) RunStateTransition(_ context.Context, ts block.TipSet, _ [][]*types.UnsignedMessage, _ [][]*types.SignedMessage, _ []block.TipSet, _ fbig.Int, _ cid.Cid, _ cid.Cid) (cid.Cid, []*types.MessageReceipt, error) {
 	stamp := ts.At(0).Timestamp
-	if pv.fullFailureTS == uint64(stamp) {
+	if pv.fullFailureTS == stamp {
 		return cid.Undef, nil, errors.New("run state transition fails on poison timestamp")
 	}
 	return cid.Undef, nil, nil
 }
 
 func (pv *poisonValidator) ValidateSemantic(_ context.Context, header *block.Block, _ block.TipSet) error {
-	if pv.headerFailureTS == uint64(header.Timestamp) {
+	if pv.headerFailureTS == header.Timestamp {
 		return errors.New("val semantic fails on poison timestamp")
 	}
 	return nil
